requirements.txt
ultralytics>=8.0.0
opencv-python
numpy
Pillow
gradio==3.50.2
